+++ src/tutorial-6-1/components/FullArticle.jsx
@@ -1,24 +1,33 @@
-import React from "react";
+import { useEffect } from "react";
 
 import { Button } from "react-bootstrap";
+import { Link, useNavigate, useParams } from "react-router-dom";
+import { posts } from "../test-data/posts";
+import { NotFound } from "../pages/404";
 
-export const FullArticle = ({ id }) => {
+export const FullArticle = () => {
+  const { id } = useParams();
+  const navigate = useNavigate();
+
+  useEffect(() => {
+    setTimeout(() => {
+      navigate("/", { replace: true });
+    }, 3000);
+  });
+
+  const currentPost = posts.find((post) => post.id === Number(id));
+
+  if (!currentPost) return <NotFound />;
+
   return (
     <div className="full-post">
-      <h1>Статья №{id}</h1>
+      <h1>{currentPost.title}</h1>
 
-      <img src="https://source.unsplash.com/400x400" alt="Article" />
-
-      <p>
-        Lorem ipsum dolor sit amet, consectetur adipisicing elit. A adipisci
-        aliquam amet asperiores aut eius minima, nemo nostrum perspiciatis
-        praesentium rerum tempora temporibus vero. Fugiat illo labore maiores
-        quam sit?
-      </p>
-
-      <a href="/">
+      <img src={currentPost.imageUrl} alt="Article" />
+      <p>{currentPost.text}</p>
+      <Link to="/">
         <Button>Назад</Button>
-      </a>
+      </Link>
     </div>
   );
 };
