+++ src/tutorial-6-1/pages/404.jsx
@@ -1,0 +1,3 @@
+export const NotFound = () => {
+  return <div>Page is not found</div>;
+};
